refactor(canvas-workspace): organize main process by domain
