+++ apps/canvas-workspace/src/main/agent/session-store.ts
