+++ apps/canvas-workspace/src/main/agent/workspace-meta.ts
