+++ apps/canvas-workspace/src/main/files/skill-installer.ts
