+++ apps/canvas-workspace/src/main/runtime/mcp-registration.ts
