+++ apps/canvas-workspace/src/main/webview/registry.ts
