+++ apps/canvas-workspace/src/plugins/main/webview-page-control/cdp-actions.ts
@@ -23,7 +23,7 @@
  * JS source — so injection is not a concern on this path.
  */
 
-import { withCdp, type CdpHost, type CdpSender } from '../../../main/cdp-session';
+import { withCdp, type CdpHost, type CdpSender } from '../../../main/webview/cdp-session';
 import { resolveKeySpec, type PageActionResult, type PageRunner } from './js-primitives';
 
 const DEFAULT_TIMEOUT_MS = 5_000;
